src/app/modulos/logistica-vehicular/detalle/@store/personal.reducer.ts
import * as actions from './personal.action';
import { IInformacionPersonalResponse } from '../@models/personal.model';

export interface State {
  loading: boolean;
  loaded: boolean;
  failed: any;
  data: IInformacionPersonalResponse;
}

export const INITIAL_STATE: State = {
  loading: false,
  loaded: false,
  failed: null,
  data: null
};

export function reducer(state = INITIAL_STATE, action: actions.Actions): State {
  if (!action) {
    return state;
  }

  switch (action.type) {
    case actions.ActionTypes.LOAD: {
      return Object.assign({}, state, {
        loading: true,
        loaded: false,
        failed: null,
        data: null
      });
    }
    case actions.ActionTypes.LOAD_SUCCESS: {
      return Object.assign({}, state, {
        loading: false,
        failed: null,
        loaded: true,
        data: action.payload
      });
    }
    case actions.ActionTypes.LOAD_FAIL: {
      return Object.assign({}, state, {
        loading: false,
        loaded: true,
        failed: action.payload
      });
    }
    default: {
      return state;
    }
  }
}

export const getData = (state: State) => state.data;
export const getLoading = (state: State) => state.loading;
export const getLoaded = (state: State) => state.loaded;
export const getFailed = (state: State) => state.failed;
